// src/services/gitlab/openapi.d.ts
export interface components {
    schemas: {
// ...
        APIEntitiesMergeRequestBasic: {
            /**
             * Format: int64
             * @example 84
             */
            id: number;
            /** @example 14 */
            iid: number;
            /**
             * Format: int64
             * @example 4
             */
            project_id?: number;
            /** @example Impedit et ut et dolores vero provident ullam est */
            title: string;
            /** @example Repellendus impedit et vel velit dignissimos. */
            description?: string | null;
            /** @example closed */
            state: string;
            /**
             * Format: date-time
             * @example 2022-08-17T12:46:35.053Z
             */
            created_at: string;
            /**
             * Format: date-time
             * @example 2022-11-14T17:22:01.470Z
             */
            updated_at?: string;
            merged_by?: components["schemas"]["APIEntitiesUserBasic"];
            merge_user?: components["schemas"]["APIEntitiesUserBasic"];
            /**
             * Format: date-time
             * @example 2022-01-31T15:10:45.080Z
             */
            merged_at?: string | null;
            closed_by?: components["schemas"]["APIEntitiesUserBasic"];
            /**
             * Format: date-time
             * @example 2022-01-31T15:10:45.080Z
             */
            closed_at?: string | null;
            title_html?: string;
            description_html?: string;
            target_branch: string;
            source_branch: string;
            user_notes_count?: number;
            upvotes?: number;
            downvotes?: number;
            author: components["schemas"]["APIEntitiesUserBasic"];
            assignee?: components["schemas"]["APIEntitiesUserBasic"];
            /** Format: int64 */
            source_project_id: number;
            /** Format: int64 */
            target_project_id: number;
            labels: string[];
            draft?: boolean;
            imported?: boolean;
            /** @example bitbucket */
            imported_from?: string;
            work_in_progress?: boolean;
            milestone?: components["schemas"]["APIEntitiesMilestone"];
            merge_when_pipeline_succeeds?: boolean;
            /** @example unchecked */
            merge_status?: string;
            /** @example mergeable */
            detailed_merge_status?: string;
            /**
             * Format: date-time
             * @example 2022-01-31T15:10:45.080Z
             */
            merge_after?: string;
            /** @example 1234abcd */
            sha: string;
            /** @example 1234abcd */
            merge_commit_sha?: string;
            /** @example 1234abcd */
            squash_commit_sha?: string;
            discussion_locked?: boolean;
            should_remove_source_branch?: boolean;
            force_remove_source_branch?: boolean;
            /**
             * Format: date-time
             * @example 2022-01-31T15:10:45.080Z
             */
            prepared_at?: string;
            allow_collaboration?: boolean;
            allow_maintainer_to_push?: boolean;
            /** @example !1 */
            reference?: string;
            references?: components["schemas"]["APIEntitiesIssuableReferences"];
            /** @example https://gitlab.example.com/my-group/my-project/-/merge_requests/1 */
            web_url: string;
            time_stats?: components["schemas"]["APIEntitiesIssuableTimeStats"];
            squash?: boolean;
            squash_on_merge?: boolean;
            task_completion_status?: components["schemas"]["APIEntitiesTaskCompletionStatus"];
            has_conflicts?: boolean;
            blocking_discussions_resolved?: boolean;
            /** @example 2 */
            approvals_before_merge?: number;
            assignees?: components["schemas"]["APIEntitiesUserBasic"][];
            reviewers?: components["schemas"]["APIEntitiesUserBasic"][];
        };
// ...
        APIEntitiesIssue: {
            /**
             * Format: int64
             * @example 84
             */
            id: number;
            /** @example 14 */
            iid: number;
            /**
             * Format: int64
             * @example 4
             */
            project_id?: number;
            /** @example Impedit et ut et dolores vero provident ullam est */
            title: string;
            /** @example Repellendus impedit et vel velit dignissimos. */
            description?: string | null;
            /** @example closed */
            state: string;
            /**
             * Format: date-time
             * @example 2022-08-17T12:46:35.053Z
             */
            created_at: string;
            /**
             * Format: date-time
             * @example 2022-11-14T17:22:01.470Z
             */
            updated_at?: string;
            /**
             * Format: date-time
             * @example 2022-11-15T08:30:55.232Z
             */
            closed_at?: string | null;
            closed_by?: components["schemas"]["APIEntitiesUserBasic"];
            /**
             * @example [
             *       "bug"
             *     ]
             */
            labels: string[];
            milestone?: components["schemas"]["APIEntitiesMilestone"];
            author: components["schemas"]["APIEntitiesUserBasic"];
            /**
             * @description One of ["ISSUE", "INCIDENT", "TEST_CASE", "REQUIREMENT", "TASK", "TICKET"]
             * @example ISSUE
             */
            type?: string;
            assignee?: components["schemas"]["APIEntitiesUserBasic"];
            user_notes_count?: number;
            merge_requests_count?: number;
            upvotes?: number;
            downvotes?: number;
            /**
             * Format: date
             * @example 2022-11-18
             */
            start_date?: string;
            /**
             * Format: date
             * @example 2022-11-20
             */
            due_date?: string;
            confidential?: boolean;
            discussion_locked?: boolean;
            /** @example issue */
            issue_type?: string;
            /** @example http://example.com/example/example/issues/14 */
            web_url: string;
            time_stats?: components["schemas"]["APIEntitiesIssuableTimeStats"];
            task_completion_status?: components["schemas"]["APIEntitiesTaskCompletionStatus"];
            /** @example 5 */
            weight?: number;
            /** @example 2 */
            blocking_issues_count?: number;
            /** @example true */
            has_tasks?: boolean;
            /** @example 2 of 4 tasks completed */
            task_status?: string;
            _links?: {
                /** @example http://example.com/api/v4/projects/1/issues/2 */
                self?: string;
                /** @example http://example.com/api/v4/projects/1/issues/2/notes */
                notes?: string;
                /** @example http://example.com/api/v4/projects/1/issues/2/award_emoji */
                award_emoji?: string;
                /** @example http://example.com/api/v4/projects/1 */
                project?: string;
                /** @example http://example.com/api/v4/projects/1/issues/75 */
                closed_as_duplicate_of?: string;
            };
            references?: components["schemas"]["APIEntitiesIssuableReferences"];
            /** @description One of ["UNKNOWN", "LOW", "MEDIUM", "HIGH", "CRITICAL"] */
            severity?: string;
            /** @example false */
            subscribed?: boolean;
            /**
             * Format: int64
             * @example 1
             */
            moved_to_id?: number;
            /** @example false */
            imported?: boolean;
            /** @example github */
            imported_from?: string;
            /** @example user@example.com */
            service_desk_reply_to?: string;
            epic_iid?: string;
            epic?: components["schemas"]["EpicBaseEntity"];
            iteration?: components["schemas"]["APIEntitiesIteration"];
            health_status?: string;
            assignees: components["schemas"]["APIEntitiesUserBasic"][];
        };
// ...
        APIEntitiesNote: {
            /**
             * Format: int64
             * @example 1
             */
            id: number;
            /** @example DiscussionNote */
            type?: string;
            /** @example Note body. */
            body: string;
            author: components["schemas"]["APIEntitiesUserBasic"];
            /**
             * Format: date-time
             * @example 2022-01-31T15:10:44.988Z
             */
            created_at: string;
            /**
             * Format: date-time
             * @example 2022-01-31T15:10:44.988Z
             */
            updated_at: string;
            /** @example false */
            system: boolean;
            /**
             * Format: int64
             * @example 1
             */
            noteable_id?: number;
            /** @example Issue */
            noteable_type?: string;
            /**
             * Format: int64
             * @example 1
             */
            project_id?: number;
            /** @example 7b09ce7e6f80347baf0316c8c94cdba9a0a7e91d */
            commit_id?: string;
            position?: Record<string, never>;
            /** @example false */
            resolvable?: boolean;
            /** @example false */
            resolved?: boolean;
            resolved_by?: components["schemas"]["APIEntitiesUserBasic"];
            /**
             * Format: date-time
             * @example 2022-01-31T15:10:44.988Z
             */
            resolved_at?: string;
            suggestions?: components["schemas"]["APIEntitiesSuggestion"];
            /** @example false */
            confidential?: boolean;
            /** @example false */
            internal?: boolean;
            /** @example false */
            imported?: boolean;
            /** @example github */
            imported_from?: string;
            /** @example 1 */
            noteable_iid?: number;
            commands_changes?: Record<string, never>;
        };
// ...
        APIEntitiesCommitStatus: {
            /**
             * Format: int64
             * @example 93
             */
            id: number;
            /** @example 18f3e63d05582537db6d183d9d557be09e1f90c8 */
            sha: string;
            /** @example develop */
            ref?: string;
            /** @example success */
            status: string;
            /** @example default */
            name: string;
            /** @example https://gitlab.example.com/janedoe/gitlab-foss/builds/91 */
            target_url?: string | null;
            description?: string | null;
            /**
             * Format: date-time
             * @example 2016-01-19T09:05:50.355Z
             */
            created_at: string;
            /**
             * Format: date-time
             * @example 2016-01-20T08:40:25.832Z
             */
            started_at?: string;
            /**
             * Format: date-time
             * @example 2016-01-21T08:40:25.832Z
             */
            finished_at?: string;
            /** @example false */
            allow_failure?: boolean;
            /**
             * Format: float
             * @example 98.29
             */
            coverage?: number;
            /**
             * Format: int64
             * @example 101
             */
            pipeline_id?: number;
            author?: components["schemas"]["APIEntitiesUserBasic"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}